+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - DatePicker:
   - Add new option `animationEnabled`
+  - Date popup in `SINGLE_DATE_SELECTED` mode, now auto close when double click
 
 ### Fixed bugs